requirements.txt
pandas
plotly
statsmodels
geopy
requests
ipykernel
black
nbformat
pyarrow
twine
scikit-learn
fsspec
gcsfs
googlemaps